+++ python_modules/dagster/dagster_tests/core_tests/engine_tests/test_child_process_executor.py
@@ -3,7 +3,7 @@ import time
 
 import pytest
 
-from dagster.core.engine.child_process_executor import (
+from dagster.core.executor.child_process_executor import (
     ChildProcessCommand,
     ChildProcessCrashException,
     ChildProcessDoneEvent,