+++ python_modules/libraries/dagster-dask/dagster_dask_tests/test_resource_tags.py
@@ -1,4 +1,4 @@
-from dagster_dask.engine import get_dask_resource_requirements
+from dagster_dask.executor import get_dask_resource_requirements
 
 from dagster import solid